refactor: improve DescribeModeledDataStreamRequest
