+++ packages/source-iotsitewise/src/asset-modules/describeModeledDataStreamRequest/index.ts
@@ -1,0 +1,2 @@
+export { DescribeModeledDataStreamRequest } from './describeModeledDataStreamRequest';
+export type { ModeledDataStream } from './types';
