+++ packages/source-iotsitewise/src/completeDataStreams.ts
@@ -4,7 +4,7 @@ import { completeAlarmStream } from './alarms/iotevents/util/completeAlarmStream
 import type { DescribeAssetModelResponse } from '@aws-sdk/client-iotsitewise';
 import type { DataStream } from '@iot-app-kit/core';
 import type { Alarms } from './alarms/iotevents';
-import type { ModeledDataStream } from './asset-modules/describeModeledDataStreamRequest/types';
+import type { ModeledDataStream } from './asset-modules/describeModeledDataStreamRequest';
 
 /**
  * Get completed data streams by merging together the data streams with the asset models and alarms.
